+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/attributes/ComposedJarRule.kt
@@ -1,0 +1,17 @@
+// Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+
+package org.jetbrains.intellij.platform.gradle.attributes
+
+import org.gradle.api.attributes.AttributeCompatibilityRule
+import org.gradle.api.attributes.CompatibilityCheckDetails
+import org.gradle.api.attributes.LibraryElements
+import org.jetbrains.intellij.platform.gradle.Constants.Configurations.Attributes
+
+abstract class ComposedJarRule : AttributeCompatibilityRule<LibraryElements> {
+
+    override fun execute(details: CompatibilityCheckDetails<LibraryElements>) = details.run {
+        if (consumerValue?.name == Attributes.COMPOSED_JAR_NAME && producerValue?.name == "jar") {
+            compatible()
+        }
+    }
+}