+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/ComposedJarTask.kt
@@ -7,11 +7,11 @@ import org.gradle.api.file.DuplicatesStrategy
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.bundling.Jar
 import org.gradle.kotlin.dsl.get
+import org.gradle.kotlin.dsl.named
 import org.jetbrains.intellij.platform.gradle.Constants.Configurations
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
 import org.jetbrains.intellij.platform.gradle.tasks.compaion.JarCompanion
 import org.jetbrains.intellij.platform.gradle.utils.extensionProvider
-import org.jetbrains.kotlin.gradle.utils.named
 
 @CacheableTask
 abstract class ComposedJarTask : Jar() {
@@ -21,9 +21,6 @@ abstract class ComposedJarTask : Jar() {
             project.registerTask<ComposedJarTask>(Tasks.COMPOSED_JAR) {
                 val jarTaskProvider = project.tasks.named<Jar>(Tasks.External.JAR)
                 val instrumentedJarTaskProvider = project.tasks.named<Jar>(Tasks.INSTRUMENTED_JAR)
-                val apiElementsConfiguration = project.configurations[Configurations.External.API_ELEMENTS]
-                val archivesConfiguration = project.configurations[Configurations.External.ARCHIVES]
-                val runtimeElementsConfiguration = project.configurations[Configurations.External.RUNTIME_ELEMENTS]
                 val intellijPlatformPluginModuleConfiguration = project.configurations[Configurations.INTELLIJ_PLATFORM_PLUGIN_MODULE]
 
                 val sourceTaskProvider = project.extensionProvider.flatMap {
@@ -48,15 +45,7 @@ abstract class ComposedJarTask : Jar() {
                 duplicatesStrategy = DuplicatesStrategy.EXCLUDE
                 JarCompanion.applyPluginManifest(this)
 
-                // Remove the default artifact exported by the current module and replace it with the final one provided by the task.
-                listOf(
-                    apiElementsConfiguration,
-                    archivesConfiguration,
-                    runtimeElementsConfiguration,
-                ).forEach { configuration ->
-                    configuration.artifacts.removeIf { it.classifier == JarCompanion.CLASSIFIER }
-                    project.artifacts.add(configuration.name, archiveFile)
-                }
+                project.artifacts.add(Configurations.INTELLIJ_PLATFORM_COMPOSED_JAR, this)
             }
     }
 }